refactor: break up the canary page into multiple files
