+++ src/components/Canary/CanaryFilters/TristateLabels.tsx
@@ -1,0 +1,65 @@
+import { useEffect, useState } from "react";
+import { TristateToggle } from "../../TristateToggle";
+import { getConciseLabelState } from "../labels";
+import { decodeUrlSearchParams, useUpdateParams } from "../url";
+
+export function TristateLabels({ labels = [] }: { labels: any[] }) {
+  const [labelStates, setLabelStates] = useState({});
+  const updateParams = useUpdateParams();
+
+  // first load or label change: set label states
+  useEffect(() => {
+    const { labels: urlLabelState = {} } = decodeUrlSearchParams(
+      window.location.search
+    );
+    const labelMap = labels.reduce((acc, current) => {
+      acc[current.id] = true;
+      return acc;
+    }, {});
+    const newLabelStates = Object.entries(urlLabelState).reduce(
+      (acc, [k, v]) => {
+        if (Object.prototype.hasOwnProperty.call(labelMap, k)) {
+          // @ts-expect-error
+          acc[k] = v;
+        }
+        return acc;
+      },
+      {}
+    );
+    setLabelStates(newLabelStates);
+  }, [labels]);
+
+  const handleToggleChange = (labelKey: string, value: string) => {
+    const { labels: urlLabelState } = decodeUrlSearchParams(
+      window.location.search
+    );
+    const newState = { ...urlLabelState };
+    newState[labelKey] = value;
+    const conciseLabelState = getConciseLabelState(newState);
+    updateParams({ labels: conciseLabelState });
+  };
+
+  return (
+    <div className="w-full break">
+      {labels
+        .filter((o) => o && o !== undefined)
+        .map((label) => (
+          <div key={label.id}>
+            <TristateToggle
+              key={label.key}
+              value={
+                Object.prototype.hasOwnProperty.call(labelStates, label.id)
+                  ? // @ts-expect-error
+                    labelStates[label.id]
+                  : 0
+              }
+              onChange={(v: string) => handleToggleChange(label.id, v)}
+              className="mb-2 flex items-center"
+              labelClass="ml-3 text-xs text-left text-gray-700 break-all overflow-ellipsis overflow-x-hidden"
+              label={label}
+            />
+          </div>
+        ))}
+    </div>
+  );
+}
